+++ sklearn/utils/_show_versions.py
